expand finetune tests to test llama
