+++ tests/sparseml/transformers/finetune/finetune_configs/gpu/gpu_config.yaml
@@ -1,0 +1,6 @@
+cadence: "nightly"
+test_type: "regression"
+model: ""zoo:llama2-7b-llama2_pretrain-base"
+dataset: wikitext
+dataset_config_name: "wikitext-2-raw-v1"
+recipe: "tests/sparseml/transformers/finetune/test_alternate_recipe.yaml"
